% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{LifeTableFertility}
\alias{LifeTableFertility}
\title{Run the LifeTableFertility Shiny App}
\usage{
LifeTableFertility()
}
\value{
Runs a Shiny app (interactive).
}
\description{
Launches the bundled Shiny application included in this package.
}
